Updated Input action maps, added input controller
We now have hopefully what should be the final update to the input action maps (Ask me if you have any problems with it). Also made an input controller that controls which input is enabled and disabled. The input controller script has a "inputActions" variable that should be abled to be called from other scripts to control what each input does. Gonna do another commit for updating the player movement script to use this.