Got basic weird shooting system
Got the playershoot script from TK, updated it to use the new input system and changed it to work with our game. Having a weird bug where the camera/characters position dictates which way the bullet goes. Also will later on add pooling so there wont be 100 of bullets spawning and despawning.