Better Shooting
The shooting mechanic is more like how its suppose to be. One issue it seems like is the camera distance and angle is messing with the mouse position numbers. Had to manually for now change the numbers so the bullet actually goes near where the mouse is. This will be upgraded later.