Added GameSceneHandler & startPoint
Added a new startPoint scene
This scene needs to be what the game points at when being compiled.
Everything else is loaded automatically
GameSceneHandler namespace also added.
This allows calls to be made directly to the handler from anywhere.
WILL be expanded on, but the current calls will not be changed without warning.