Added New Feature: UI FadeIn/Out Effects
Can be inherited by and given additional functionality. Good for context sensitivity.
The class will automatically keep track of all original value and will return everything to the state it found it in.

Updated the plant health bar controller, and the hub controller with cool new context sensitive feature.
They only show up when they're not at full health.
There is a toggle on the health bar to turn it off.
Also integrated the Hub health bar with a damage controller, as well as made an event whenever it gets hit.
Also fixed the Seed dropper so it won't crash if it's empty.