Added New Effect: Damage Flash
New MonoBehavior that can be added to any object with a sprite renderer that'll be able to make it flicker with a white color.
It indexes all the childen and tracks their independant original materials.
Tested  on the player
(If you wanna see then just copy it from onHit to onAttack)