Added EnemySpawnList
There is now a static script called EnemySpawnList where you can get the transform of all spawnpoints. This can be used to get the locations of each. Also made the spawn areas wider and enemies now spawn more when all waves are completed, this can be changed later to be when the next night happens.